Components placed.  Needs routing.  Happy Hanukkah!
There's really not much more to say on this one, but I'd hate to
dissapoint the diligent fellow that actually reads these notes.  I will
add that there is plenty of routing to be done, and I just may try the
autorouter for kicks.  Also, congnac and cigars make for cheerful vices.
If I have to burn hitpoints for pleasure, man that's the way to fly!