Deleted ZIP files of 3D connectors and just left STP files.
Re-discovered the 5/32" aluminum nylock that we purchased and settled on #2-56 hardware because we have it and it's easy. Moved hardware as necessary to make sure things fit... width now set by OD of FDM'd standoffs on #2 bolts, which is funny. Now ready to hand off to Nathan for 3D modeling.